media: vcodec: remove useless type convert
remove useless type convert

CR-Id: ALPS04383538
Feature: Video Framework

+++ drivers/media/platform/mtk-vcodec/venc_drv_if.c
@@ -77,7 +77,7 @@ int venc_if_get_param(struct mtk_vcodec_ctx *ctx, enum venc_get_param_type type,
 	if (!ctx->drv_handle) {
 		inst = kzalloc(sizeof(struct venc_inst), GFP_KERNEL);
 		inst->ctx = ctx;
-		ctx->drv_handle = (void *)(unsigned long)(inst);
+		ctx->drv_handle = (void *)(inst);
 		#ifdef CONFIG_VIDEO_MEDIATEK_VCU
 		ctx->enc_if = get_enc_common_if();
 		#endif
